Broadway Tickets: Lucky
Most of my friends in New York are expected to list "Go to Broadway to watch a musical" as one of the must-have items. This app helps you participate in Broadway tickets, with periodic draws and daily draws. Many repertoires are waiting for you to choose, for example, Aladdin Lion King Charlie and the Chocolate Factory.

Features:
easy to use
Simple interface